Split memory-circuit kernels into plain and realtime families
